# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Fleet EDR API
  version: "0.1.0"
  description: |
    HTTP API for Fleet EDR. Three audience tiers:

    - **Agents** (macOS endpoints) authenticate with a per-host bearer
      token obtained at enrollment. They call `POST /api/enroll`,
      `POST /api/events`, and the host-scoped `/commands` endpoints.
    - **Browsers** (admin UI) authenticate with an `edr_session` cookie
      (HttpOnly, Secure, SameSite=Lax) plus an `X-CSRF-Token` header on
      unsafe methods. They call the session-protected endpoints.
    - **Operators** hit `/livez`, `/readyz`, `/health` with no auth.

    For prose context see
    [docs/api.md](https://github.com/getvictor/fleet-edr/blob/main/docs/api.md).
    For a human-readable host/agent install guide see
    [docs/install-agent-manual.md](https://github.com/getvictor/fleet-edr/blob/main/docs/install-agent-manual.md).
  license:
    name: Fleet EDR License
    url: https://github.com/getvictor/fleet-edr/blob/main/LICENSE
servers:
  - url: https://edr.example.com
    description: Your EDR server
...
paths:
...
  /api/hosts:
    get:
      tags: [Hosts]
      summary: List hosts
      security:
        - SessionCookie: []
      responses:
        "200":
          description: Host summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostSummary"
  /api/hosts/{host_id}/tree:
    get:
      tags: [Hosts]
      summary: Process tree for a host
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/HostID"
        - name: from
          in: query
          schema: {type: integer, format: int64}
          description: Lower bound, ns since epoch. Default `now - 1h`.
        - name: to
          in: query
          schema: {type: integer, format: int64}
          description: Upper bound, ns since epoch. Default `now`.
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 5000, default: 2000}
      responses:
        "200":
          description: Forest of process roots
          content:
            application/json:
              schema:
                type: object
                properties:
                  roots:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProcessNode"
  /api/hosts/{host_id}/processes/{pid}:
    get:
      tags: [Hosts]
      summary: Process detail
      security:
        - SessionCookie: []
      parameters:
        - $ref: "#/components/parameters/HostID"
        - name: pid
          in: path
          required: true
          schema: {type: integer}
        - name: at
          in: query
          schema: {type: integer, format: int64}
          description: Point-in-time ns since epoch. Default `now`.
      responses:
        "200":
          description: Process detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Process"
        "404":
          description: No process found at the given coordinates
  /api/alerts:
    get:
      tags: [Alerts]
      summary: List alerts
      security:
        - SessionCookie: []
      parameters:
        - name: host_id
          in: query
          schema: {type: string}
        - name: status
          in: query
          schema:
            type: string
            enum: [open, acknowledged, resolved]
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: process_id
          in: query
          schema: {type: integer, format: int64}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, default: 100}
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /api/alerts/{id}:
    get:
      tags: [Alerts]
      summary: Alert detail (with event IDs)
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        "200":
          description: Alert with linked events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertDetail"
        "404":
          description: Alert not found
    put:
      tags: [Alerts]
      summary: Update alert status
      security:
        - SessionCookie: []
          CSRF: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [open, acknowledged, resolved]
      responses:
        "204":
          description: Status updated
        "400":
          description: Invalid status value
        "404":
          description: Alert not found
  /api/commands:
    get:
      tags: [Commands]
      summary: List commands for the authenticated host
      description: |
        Host-token-only. The `host_id` is always the authenticated host's
        id; any query parameter is ignored. Polling this endpoint also
        refreshes the host's `last_seen` timestamp.
      security:
        - HostToken: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, acked, completed, failed]
      responses:
        "200":
          description: Command list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Command"
    post:
      tags: [Commands]
      summary: Enqueue a command for a host (admin)
      security:
        - SessionCookie: []
          CSRF: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host_id, command_type]
              properties:
                host_id: {type: string}
                command_type: {type: string}
                payload: {type: object, additionalProperties: true}
      responses:
        "201":
          description: Command created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: {type: integer, format: int64}
  /api/commands/{id}:
    get:
      tags: [Commands]
      summary: Get a single command (admin)
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        "200":
          description: Command
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Command"
        "404":
          description: Not found
    put:
      tags: [Commands]
      summary: Update command status (agent)
      description: |
        Agents call this with `status=acked` once they pick up a command
        and with `status=completed` or `status=failed` once they finish.
        The handler enforces that the command belongs to the authenticated
        host — agent A cannot touch agent B's commands.
      security:
        - HostToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [acked, completed, failed]
                result:
                  type: object
                  description: Command-type-specific result blob, free-form JSON
      responses:
        "204":
          description: Status updated
        "400":
          description: Invalid status
        "404":
          description: Command not found (or not owned by authenticated host)
...
components:
  securitySchemes:
    HostToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Per-host bearer token issued by `POST /api/enroll`. Pass as
        `Authorization: Bearer <host_token>`.
    SessionCookie:
      type: apiKey
      in: cookie
      name: edr_session
      description: |
        HttpOnly cookie set by `POST /api/session`. The browser
        attaches it automatically.
    CSRF:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Required on unsafe methods (`POST`, `PUT`, `DELETE`) for
        session-cookie-authenticated endpoints. Value comes from the
        login response's `csrf_token` field.
  parameters:
    HostID:
      name: host_id
      in: path
      required: true
      description: Host identifier (UUID-shaped opaque string)
      schema: {type: string}
  schemas:
...
    HostSummary:
      type: object
      properties:
        host_id: {type: string}
        event_count: {type: integer, format: int64}
        last_seen_ns: {type: integer, format: int64}
    Process:
      type: object
      properties:
        id: {type: integer, format: int64}
        host_id: {type: string}
        pid: {type: integer}
        ppid: {type: integer}
        path: {type: string}
        args: {type: array, items: {type: string}}
        uid: {type: integer}
        gid: {type: integer}
        code_signing:
          type: object
          properties:
            team_id: {type: string}
            signing_id: {type: string}
            flags: {type: integer}
            is_platform_binary: {type: boolean}
        sha256: {type: string}
        fork_time_ns: {type: integer, format: int64}
        exec_time_ns: {type: integer, format: int64}
        exit_time_ns: {type: integer, format: int64}
        exit_code: {type: integer}
    ProcessNode:
      allOf:
        - $ref: "#/components/schemas/Process"
        - type: object
          properties:
            children:
              type: array
              items:
                $ref: "#/components/schemas/ProcessNode"
    Alert:
      type: object
      required: [id, host_id, rule_id, severity, title, process_id, status, created_at, updated_at]
      properties:
        id: {type: integer, format: int64}
        host_id: {type: string}
        rule_id: {type: string}
        severity: {type: string, enum: [low, medium, high, critical]}
        title: {type: string}
        description: {type: string}
        process_id: {type: integer, format: int64}
        status: {type: string, enum: [open, acknowledged, resolved]}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        resolved_at: {type: string, format: date-time}
    AlertDetail:
      allOf:
        - $ref: "#/components/schemas/Alert"
        - type: object
          properties:
            event_ids:
              type: array
              items: {type: string, format: uuid}
    Command:
      type: object
      required: [id, host_id, command_type, status, created_at]
      properties:
        id: {type: integer, format: int64}
        host_id: {type: string}
        command_type: {type: string}
        payload:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [queued, acked, completed, failed]
        created_at: {type: string, format: date-time}
        acked_at: {type: string, format: date-time}
        completed_at: {type: string, format: date-time}
        result:
          type: object
          additionalProperties: true
...